eventos/requirements.txt
flask==1.1.1
google-cloud-storage==1.19.1
google-cloud-error-reporting==0.33.0

